main camera as observation
