+++ metadrive/engine/base_engine.py
@@ -422,7 +422,7 @@ class BaseEngine(EngineCore, Randomizable):
             return self.replay_manager.replay_agents
 
     def setup_main_camera(self):
-        from metadrive.engine.core.chase_camera import MainCamera
+        from metadrive.engine.core.main_camera import MainCamera
         if self.global_config["use_render"] or (self.global_config["offscreen_render"] and
                                                 self.global_config["vehicle_config"]["image_source"] == "main_camera"):
             return MainCamera(self, self.global_config["camera_height"], self.global_config["camera_dist"])
